Add fixes after review
